+++ pyproject.toml
@@ -36,8 +36,9 @@ cmake = ">=3.25.0"
 pygithub = ">=1.58.2"
 githubrelease = ">=1.5.9"
 httpx = ">=0.18.2"
-autopub = ">=0.2.2"
+autopub = "0.3.0"
 mimesis = ">=10.1.0"
+pytest = "7.3.1"
 pytest-cov = ">=4.1.0"
